project restructure to use gradle building process for intellij. Moved project structure to reflect gradle build process
