+++ build.gradle
@@ -1,0 +1,42 @@
+plugins {
+    id 'java'
+    id "org.openjfx.javafxplugin" version "0.0.9"
+    id 'idea'
+    id 'application'
+}
+group 'MenuGO'
+version '0.1'
+
+mainClassName = "Main" // replace with your main class
+
+
+
+javafx {
+    modules = [ 'javafx.controls', 'javafx.fxml', 'javafx.graphics', 'javafx.base']
+}
+
+group 'org.example'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+sourceSets {
+    main {
+        java {
+            srcDirs = ['src']
+        }
+    }
+}
+dependencies {
+    testCompile(
+            [group: 'junit', name: 'junit', version: '4.12'],
+            [group: 'org.xerial', name: 'sqlite-jdbc', version: '3.27.2']
+
+    )
+    compile(
+            [group: 'junit', name: 'junit', version: '4.12'],
+            [group: 'org.xerial', name: 'sqlite-jdbc', version: '3.27.2']
+    )
+}
+
