--- src/DBAccess/MenuAccess.java
@@ -1,4 +1,0 @@
-package DBAccess;
-
-public class MenuAccess extends  DBAccess{
-}
