--- src/Models/Ingredient.java
@@ -1,28 +1,0 @@
-package Models;
-
-public class Ingredient {
-
-    private String name;
-    private String category;
-
-    public Ingredient(String name, String category) {
-        this.name = name;
-        this.category = category;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCategory() {
-        return category;
-    }
-
-    public void setCategory(String category) {
-        this.category = category;
-    }
-}
