+++ src/main/java/MenuGO/DBAccess/MenuAccess.java
@@ -1,0 +1,4 @@
+package MenuGO.DBAccess;
+
+public class MenuAccess extends  DBAccess{
+}
