+++ src/main/java/MenuGO/Models/MealItem.java
@@ -1,0 +1,29 @@
+package MenuGO.Models;
+
+public class MealItem {
+    private Ingredient ingredient;
+    private Quantity quantity;
+
+    public MealItem(Ingredient ingredient, Quantity quantity) {
+        this.ingredient = ingredient;
+        this.quantity = quantity;
+    }
+
+    public Ingredient getIngredient() {
+        return ingredient;
+    }
+
+    public String getIngredientName() { return ingredient.getName();}
+
+    public void setIngredient(Ingredient ingredient) {
+        this.ingredient = ingredient;
+    }
+
+    public Quantity getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(Quantity quantity) {
+        this.quantity = quantity;
+    }
+}
