+++ src/main/java/MenuGO/Models/Menu.java
@@ -1,0 +1,18 @@
+package MenuGO.Models;
+
+import java.util.ArrayList;
+import java.util.Date;
+
+public class Menu {
+    ArrayList<Meal> meals;
+    Date startDate;
+    Date endDate;
+
+    public Menu(ArrayList<Meal> meals, Date startDate, Date endDate) {
+        this.meals = meals;
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+
+
+}
