+++ src/main/resources/ListOfMeals.fxml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="MenuGO.Controllers.ListOfMealsController">
+   <children>
+      <ListView layoutX="91.0" layoutY="186.0" prefHeight="200.0" prefWidth="200.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="50.0" />
+      <Button fx:id="addNewMeal" layoutX="274.0" layoutY="25.0" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="50.0" text="Add New Meal" AnchorPane.leftAnchor="220.0" AnchorPane.rightAnchor="220.0" AnchorPane.topAnchor="15.0" />
+      <Button fx:id="home" layoutX="14.0" layoutY="15.0" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="50.0" text="home" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="15.0" />
+   </children>
+</AnchorPane>
