--- src/sample/CreateMenu.fxml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controllers.CreateMenuController">
-   <children>
-      <Button fx:id="home" layoutX="14.0" layoutY="14.0" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="50.0" text="home" AnchorPane.leftAnchor="15.0" AnchorPane.topAnchor="15.0" />
-   </children>
-</AnchorPane>
