--- src/sample/sample.fxml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-
-<GridPane alignment="center" hgap="10" prefHeight="241.0" prefWidth="648.0" vgap="10" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.Controller">
-   <columnConstraints>
-      <ColumnConstraints />
-      <ColumnConstraints />
-      <ColumnConstraints />
-      <ColumnConstraints />
-      <ColumnConstraints />
-      <ColumnConstraints />
-   </columnConstraints>
-   <rowConstraints>
-      <RowConstraints />
-      <RowConstraints />
-      <RowConstraints />
-      <RowConstraints />
-   </rowConstraints>
-   <children>
-      <Button fx:id="createMenu" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="140.0" text="Create a Menu" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-      <Button fx:id="viewMealList" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="140.0" text="View Menu List" GridPane.columnIndex="3" GridPane.rowIndex="1" />
-      <Button fx:id="createGroceryList" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="140.0" text="Generate Grocery List" GridPane.columnIndex="5" GridPane.rowIndex="1" />
-      <Button fx:id="lazyFeature" cache="true" mnemonicParsing="false" onAction="#buttonPressedEvent" prefHeight="26.0" prefWidth="462.0" text="I wAnT tO bE lAzY / aDvEnTuRoUs" textAlignment="JUSTIFY" GridPane.columnIndex="1" GridPane.columnSpan="5" GridPane.rowIndex="3" />
-   </children>
-</GridPane>
